+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Publish process fails without retry in three specified situations. Bug in lookup
 dev&pre env add for DCC in configClient.properties, and several bug fixes.
 ### Features:
 + dev env appended in nested configClient.properties.[#47](http://gitlab.qima-inc.com/paas/nsq-client-java/issues/47)
++ publish API with shardingID as parameter.[#49](http://gitlab.qima-inc.com/paas/nsq-client-java/issues/49)
 ### Fixes:
 + synchronization in boostrap map in producer connection pool factory.[#46](http://gitlab.qima-inc.com/paas/nsq-client-java/issues/46)
 + nsq.sdk.env not work when set as system property  [#48](http://gitlab.qima-inc.com/paas/nsq-client-java/issues/48)